"Update constants and components for clothes and home appliances footprints; add home appliances component"
This diff includes updates to constants and components related to the calculation of footprints for clothes and home appliances. It replaces the use of `CLOTHES_FACTORS` with `CLOTHES_FOOTPRINTS` and adds a new `HOME_APPLIANCE_FOOTPRINTS` object to constants. The `getHomeApplianceFootprint` function has been updated to use these new footprints and the `getAge` function has been added to determine the age of a home appliance. The `HomeAppliances` component has been added to the `Divers` component and will calculate and display the footprint for home appliances. The `fr.json` file has also been updated to include translations for the new home appliances section.